pass MAKE env. var. to clisp-link, so ordrules/link.sh will use the proper make.
without this, there was a problem e.g. on FreeBSD when using gmake to build xindy,
which was setting MAKEFLAGS=w but ordrules/link.sh was calling "make" (because MAKE was not passed to it),
and "make" was breaking because of the MAKEFLAGS variable, thinking it is a target to make.