fit auto-detection of installation options
install-tl currently checks for the presence of the directory
	texmf-dist/web2c
when determining whether there is an uncompressed installation
available. This breaks since now this directory is present even
in the network installer.

Change it to check for readability of texmf-dist/web2c/texmf.cnf
which is not available in the installer.